1. 添加对UNIX的测试 2. 测试搜索相关的功能

+++ src/ali_search/res_value.cpp
@@ -70,8 +70,8 @@ namespace aos{
     }
     for (std::size_t i = 0; i < error_res.size(); i++){
       AosError aos_error;
-      aos_error.error_code = error_res[i][JSON_ERRORS_CODE].asUInt();
-      aos_error.message = error_res[i][JSON_ERRORS_MESSAGE].asString();
+      aos_error.error_code = error_res[(int)i][JSON_ERRORS_CODE].asUInt();
+      aos_error.message = error_res[(int)i][JSON_ERRORS_MESSAGE].asString();
       aos_errors_.push_back(aos_error);
       LOG_ERROR << "[ " << aos_error.error_code << " ] "
         << aos_error.message;
